ContactUsRatingViewModelTest bug fix

+++ features/operational/contact_us/src/test/java/com/tokopedia/contactus/inboxticket2/view/presenter/ContactUsRatingViewModelTest.kt
@@ -170,7 +170,7 @@ class ContactUsRatingViewModelTest {
         val badCsatReasonListItem: ArrayList<BadCsatReasonListItem> = arrayListOf()
         for (i in 1..6) {
             badCsatReasonListItem.add(BadCsatReasonListItem().apply {
-                id = i
+                id = i.toLong()
                 messageEn = "message-$i"
                 message = "message-$i"
             })
